fixed the functions with data re-formatting of data from 'noos_read.m'
Problem was that the noos_read.m provides data in a different format if you request information for only one location (i.e. method 2 of matroos_get_series_example.m). Solution to add a check of the format of the h,t an v variables and re-cast them if necessary.

svn path=/trunk/matlab/; revision=7774


Former-commit-id: 005db18870eb9653292ac5199b78425be267330a