End points are now consistently kept in the ldb when shortening.
e.g: cutting up an ldb part of 190 mtr. in 100 mtr sections would result in 1 section of 100 mtr., now it results 100 and 90 mtr.

Future: make a uniform cut-up with max. value dx, so for the above example result in 95 and 95 mtr. parts.

help description added along with simple UI when not specifying input variables..

svn path=/trunk/matlab/; revision=10568


Former-commit-id: 80eb2565549927fa85e3273c5418b69cca6729c7