added functionality for equidistant dx cutting, while keeping default functionality for DDB use, it is now possible to (optionally) add 'extact' or 'equi' to the input, when using 'equi', the problem described below (after 'Future') is solved. Cutting up an ldb of 100 mtr in dx = 40 sections results in 0 - 40 - 80 - 100 but when using 'equi' it results in 0 - 33.3 - 66.7 - 100 (handy for unibest, and thus newly implemented in writeMDA).
freek.scheel@deltares.nl

+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++

End points are now consistently kept in the ldb when shortening.

e.g: cutting up an ldb part of 190 mtr. in 100 mtr sections would result in 1 section of 100 mtr., now it results 100 and 90 mtr.

Future: make a uniform cut-up with max. value dx, so for the above example result in 95 and 95 mtr. parts.

help description added along with simple UI when not specifying input variables..

svn path=/trunk/matlab/; revision=10582


Former-commit-id: c49f872c877c502d21c35863f0ab8b266b17785c