Updated the ldbTool:
New plotting options:

- Show dots on landboundary points
- Show indices (numbers) on landboundary points

Some fixes for versions of Matlab using the new plotting engine (since 2015):

- Filled landboundaries can be removed again
- Start & end points can be removed again

When loading a landboundary that starts or ends with no-data values, these are removed when saving to file again.
The user is now warned when loading a landboundary file which will be changed when loaded.

Freek Scheel

svn path=/trunk/matlab/; revision=14281


Former-commit-id: 4d5b252b8d8fdd7a73ff8bd4f0097e776dccaa91